refactoring: moved code to api_converter.py code cleanup
