+++ services/web/server/src/server/comp_backend_api.py
@@ -15,13 +15,12 @@ from aiohttp import web, web_exceptions
 
 import sqlalchemy.exc
 from s3wrapper.s3_client import S3Client
-from simcore_director_sdk.rest import ApiException
 from simcore_sdk.models.pipeline_models import (Base, ComputationalPipeline,
                                                 ComputationalTask)
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from . import director_sdk
+from . import api_converter
 from .comp_backend_worker import celery
 
 # TODO: this should be coordinated with postgres options from config/server.yaml
@@ -83,59 +82,74 @@ async def async_request(method, session, url, data=None, timeout=10):
                 return await response.json()
         # TODO: else raise ValueError method not implemented?
 
-async def __convert_task_to_old_version(task):
-    _LOGGER.debug("converting task to old version %s", task)
-    image_key = task["image"]
-    try:
-        services_enveloped = await director_sdk.get_director().services_by_key_version_get(image_key["name"], image_key["tag"])
-    except ApiException as err:
-        _LOGGER.exception("Error while converting to old version:")
-        raise web_exceptions.HTTPNotFound(reason=str(err))
-    if not services_enveloped:
-        raise web_exceptions.HTTPNotFound(reason="Could not retrieve service")
-    node_details = services_enveloped.data[0]
-    _LOGGER.debug("retrieved service details: %s", node_details)
-    # let's convert
-    old_task = task
-    old_task["input"] = __convert_ports_to_old_version(task["input"], node_details.inputs)
-    old_task["output"] = __convert_ports_to_old_version(task["output"], node_details.outputs)
-    _LOGGER.debug("Completed conversion:%s", old_task)
-    return old_task
+async def _parse_pipeline(pipeline_data):
+    dag_adjacency_list = dict()
+    tasks = dict()
+    io_files = []
+    
+    for node_uuid, node_data in pipeline_data.items():
+        node_key = node_data["key"]
+        node_version = node_data["version"]
+        node_inputs = node_data["inputs"]
+        node_outputs = node_data["outputs"]
+        _LOGGER.debug("node %s:%s has inputs: \n%s\n outputs: \n%s", node_key, node_version, node_inputs, node_outputs)
+        #TODO: we should validate all these things before processing...
 
-def __convert_ports_to_old_version(ports, node_description):
-    _LOGGER.debug("converting ports from %s using description %s", ports, node_description)
-    old_ports = []
-    for port_key, port_data in ports.items():
-        # sanitize port key (some old services define some keys with a . inside that is not accepted by the schema)
-        #TODO: Remove this once the services are updated
-        for key in node_description.keys():
-            if str(key).startswith(port_key):
-                port_key = key
-                break
-        # port_key = str(port_key).partition(".")[0]
-        port_description = node_description[port_key]
-        old_port = {
-            "key":port_key,
-            "label":port_description["label"],
-            "desc":port_description["description"],
-            "type":port_description["type"]
+        # build adjacency list
+        for input_data in node_inputs.values():
+            if not isinstance(input_data, dict):
+                continue
+            if "nodeUuid" in input_data and "output" in input_data:
+                input_node_uuid = input_data["nodeUuid"]
+
+                if input_node_uuid not in dag_adjacency_list:
+                    dag_adjacency_list[input_node_uuid] = []
+                if node_uuid not in dag_adjacency_list[input_node_uuid]:
+                    dag_adjacency_list[input_node_uuid].append(node_uuid)
+            
+        for output_key, output_data in node_outputs.items():
+            if not isinstance(output_data, dict):
+                continue
+            if "store" in output_data and "path" in output_data:
+                if output_data["store"] == "s3-z43":
+                    current_filename_on_s3 = output_data["path"]
+                    if current_filename_on_s3:
+                        new_filename = node_uuid + "/" + output_key # in_1
+                        # copy the file
+                        io_files.append({ "from" : current_filename_on_s3, "to" : new_filename })
+
+        # create the task
+        task = {
+            "input":node_inputs,
+            "output":node_outputs,
+            "image":{
+                "name":node_key,
+                "tag":node_version
+            }
         }
-        if port_description["type"] == "data:*/*":
-            old_port["type"] = "file-url"
-        elif port_description["type"] == "data:application/zip":
-            old_port["type"] = "folder-url"
-        
-        if port_data:
-            old_port["value"] = port_data
-            if isinstance(port_data, dict):
-                old_port["value"] = "null"
-                if "nodeUuid" in port_data and "output" in port_data:
-                    old_port["value"] = str(".").join(["link", port_data["nodeUuid"], port_data["output"]])
 
-        else:
-            old_port["value"] = "null"
-        old_ports.append(old_port)
-    return old_ports
+        # currently here a special case to handle the built-in file manager that should not be set as a task
+        if str(node_key).count("FileManager") > 0:
+            continue
+
+        # TODO: SAN This is temporary. As soon as the services are converted this should be removed.
+        task = await api_converter.convert_task_to_old_version(task)
+
+        tasks[node_uuid] = task
+
+    return dag_adjacency_list, tasks, io_files
+
+async def _transfer_data(app, pipeline_id, io_files):
+    if io_files:
+        _config = app["config"]["s3"]
+
+        s3_client = S3Client(endpoint=_config['endpoint'], access_key=_config['access_key'], secret_key=_config['secret_key'])
+        for io_file in io_files:
+            _from = io_file["from"]
+            _to = str(pipeline_id) + "/" + io_file["to"]
+            _LOGGER.debug("COPYING from %s to %s", _from, _to )
+            #TODO: make async?
+            s3_client.copy_object(_config['bucket_name'], _to, _from)
 
 # pylint:disable=too-many-branches, too-many-statements
 @comp_backend_routes.post("/start_pipeline")
@@ -157,160 +171,31 @@ async def start_pipeline(request):
     """
     # FIXME: this should be implemented generaly using async lazy initialization of db_session??
     #pylint: disable=W0603
-
-    
     global db_session
     if db_session is None:
         await init_database(request.app)
-
     request_data = await request.json()
 
     _LOGGER.debug("Client calls start_pipeline with %s", request_data)
-
     _app = request.app["config"]
-    response = {}
-
-    dag_adjacency_list = dict()
-    tasks = dict()
-
-    pipeline_id = -1
-    pipeline_name = "New pipeline"
-    _LOGGER.debug("Start Pipeline")
-
-    io_files = []
-    pipeline = request_data["workbench"]
-    for node_uuid, node_data in pipeline.items():
-        _LOGGER.debug("node: %s:%s", node_uuid, node_data)
-
-        node_key = node_data["key"]
-        node_version = node_data["version"]
-        node_inputs = node_data["inputs"]
-        node_outputs = node_data["outputs"]
-
-        _LOGGER.debug("node %s:%s has inputs: %s\n outputs: %s", node_key, node_version, node_inputs, node_outputs)
-
-        # node linking
-        for input_key, input_data in node_inputs.items():
-            if not isinstance(input_data, dict):
-                continue
-            if "nodeUuid" in input_data and "output" in input_data:
-                input_node_uuid = input_data["nodeUuid"]
-                input_node_port_key = input_data["output"]
-
-                if input_node_uuid not in dag_adjacency_list:
-                    dag_adjacency_list[input_node_uuid] = []
-                if node_uuid not in dag_adjacency_list[input_node_uuid]:
-                    dag_adjacency_list[input_node_uuid].append(node_uuid)
-            
-        for output_key, output_data in node_outputs.items():
-            if not isinstance(output_data, dict):
-                continue
-            
-            if "store" in output_data and "path" in output_data:
-                if output_data["store"] is "z43":
-                    current_filename_on_s3 = output_data["path"]
-                    if current_filename_on_s3:
-                        new_filename = node_uuid + "/" + output_key # in_1
-                        # copy the file
-                        io_files.append({ "from" : current_filename_on_s3, "to" : new_filename })
-
-        # create the task
-        task = {
-            "input":node_inputs,
-            "output":node_outputs,
-            "image":{
-                "name":node_key,
-                "tag":node_version
-            }
-        }
-
-        if str(node_key).count("FileManager") > 0:
-            continue
-
-        if True:
-            task = await __convert_task_to_old_version(task)
-            
-
-        tasks[node_uuid] = task
-
-    # links = request_data["links"]
-    # # pylint: disable=too-many-nested-blocks
-    # try:
-    #     io_files = []
-    #     for node in nodes:
-    #         _LOGGER.debug("NODE %s ", node)
-
-    #         node_id = node["uuid"]
-    #         # find connections
-    #         successor_nodes = []
-    #         task = {}
-    #         is_io_node = False
-    #         if node["key"] == "FileManager":
-    #             is_io_node = True
-
-    #         task["input"] = node["inputs"]
-    #         task["output"] = node["outputs"]
-    #         task["image"] = {"name" : node["key"], "tag"  : node["tag"]}
-
-    #         if is_io_node:
-    #             for ofile in node["outputs"]:
-    #                 current_filename_on_s3 = ofile["value"]
-    #                 if current_filename_on_s3:
-    #                     new_filename = node_id +"/" + ofile["key"] # out_1
-    #                     # copy the file
-    #                     io_files.append({ "from" : current_filename_on_s3, "to" : new_filename })
-
-    #         for link in links:
-    #             if link["node1Id"] == node_id:
-    #                 successor_node_id = link["node2Id"]
-    #                 if successor_node_id not in successor_nodes and not is_io_node:
-    #                     successor_nodes.append(successor_node_id)
-    #             if link["node2Id"] == node_id:
-    #                 # there might be something coming in
-    #                 predecessor_node_id = link["node1Id"]
-    #                 output_port = link["port1Id"]
-    #                 input_port = link["port2Id"]
-    #                 # we use predecessor_node_id.output_port as id fo the input
-    #                 for t in task["input"]:
-    #                     if t["key"] == input_port:
-    #                         t["value"] = "link." + predecessor_node_id + "." + output_port
-
-    #         if not is_io_node:
-    #             # a node can have an empty successor
-    #             #if len(successor_nodes):
-    #             dag_adjacency_list[node_id] = successor_nodes
-    #             tasks[node_id] = task
-
+    _LOGGER.debug("Parse pipeline")
+    dag_adjacency_list, tasks, io_files = await _parse_pipeline(request_data["workbench"])
     _LOGGER.debug("Pipeline parsed")
     try:
+        # create the new pipeline in db
         pipeline = ComputationalPipeline(dag_adjacency_list=dag_adjacency_list, state=0)
         _LOGGER.debug("Pipeline object created")
-
         db_session.add(pipeline)
         _LOGGER.debug("Pipeline object added")
-
         db_session.flush()
         _LOGGER.debug("Pipeline flushed")
-
         pipeline_id = pipeline.pipeline_id
 
         # now we know the id, lets copy over data
-        if io_files:
-            _config = _app["config"]["s3"]
-
-            s3_client = S3Client(endpoint=_config['endpoint'],
-                access_key=_config['access_key'], secret_key=_config['secret_key'])
-            for io_file in io_files:
-                _from = io_file["from"]
-                _to = str(pipeline_id) + "/" + io_file["to"]
-                _LOGGER.debug("COPYING from %s to %s", _from, _to )
-                #TODO: make async?
-                s3_client.copy_object(_config['bucket_name'], _to, _from)
-
-
+        await _transfer_data(_app, pipeline_id, io_files)
+        # create the tasks in db
         pipeline_name = "request_data"
         internal_id = 1
-
         for node_id in tasks:
             task = tasks[node_id]
             new_task = ComputationalTask( \
@@ -324,29 +209,28 @@ async def start_pipeline(request):
                 )
             internal_id = internal_id+1
             db_session.add(new_task)
-
         db_session.commit()
-
+        # commit the tasks to celery
         task = celery.send_task("comp.task", args=(pipeline_id,), kwargs={})
         _LOGGER.debug("Task commited")
-
-    except sqlalchemy.exc.SQLAlchemyError as err:
-        _LOGGER.exception("Alchemy error. Rolling backe and returning pipeline_id=-1")
+        # answer the client
+        response = {
+        "pipeline_name":pipeline_name,
+        "pipeline_id":str(pipeline_id)
+        }
+        _LOGGER.debug("END OF ROUTINE. Response %s", response)
+        return web.json_response(response, status=201)
+    except sqlalchemy.exc.InvalidRequestError as err:
+        _LOGGER.exception("Alchemy error: Invalid request. Rolling back db.")
         db_session.rollback()
-        pipeline_id = -1
         raise web_exceptions.HTTPInternalServerError(reason=str(err)) from err
-
+    except sqlalchemy.exc.SQLAlchemyError as err:
+        _LOGGER.exception("Alchemy error: General error. Rolling back db.")
+        db_session.rollback()
+        raise web_exceptions.HTTPInternalServerError(reason=str(err)) from err
     except Exception as err:
-        _LOGGER.exception("Unexpected Exception. Returning pipeline_id=-1")
-        pipeline_id = -1
+        _LOGGER.exception("Unexpected error.")
         raise web_exceptions.HTTPInternalServerError(reason=str(err)) from err
-
     finally:
         _LOGGER.debug("Close session")
         db_session.close()
-
-    _LOGGER.debug("END OF ROUTINE. Response %s", response)
-    response['pipeline_name'] = pipeline_name
-    response['pipeline_id'] = str(pipeline_id)
-
-    return web.json_response(response)
